Resolving form conflict by creating FormCard component.
