+++ src/components/Form/FormFooter.react.js
@@ -1,0 +1,16 @@
+// @flow
+
+import * as React from "react";
+import cn from "classnames";
+
+type Props = {|
+  +children?: React.Node,
+  +className?: string,
+|};
+
+function FormFooter(props: Props): React.Node {
+  const classes = cn("form-footer", props.className);
+  return <div className={classes}>{props.children}</div>;
+}
+
+export default FormFooter;
