--- src/forms/Form.react.js
@@ -1,29 +1,0 @@
-// @flow
-
-import * as React from "react";
-
-type Props = {|
-  +action: string,
-  +children?: React.Node,
-  +method: string,
-  +title: string,
-  +buttonText: string,
-|};
-
-function Form(props: Props): React.Node {
-  return (
-    <form className="card" action={props.action} method={props.method}>
-      <div className="card-body p-6">
-        <div className="card-title">{props.title}</div>
-        {props.children}
-        <div className="form-footer">
-          <button type="submit" className="btn btn-primary btn-block">
-            {props.buttonText}
-          </button>
-        </div>
-      </div>
-    </form>
-  );
-}
-
-export default Form;
