+++ src/index.js
@@ -7,6 +7,6 @@ export {
 export {
   default as DefaultErrorPage,
 } from "./page_templates/DefaultErrorPage.react";
-export { default as Form } from "./forms/Form.react";
+export { default as FormCard } from "./forms/FormCard.react";
 export { default as FormTextInput } from "./forms/FormTextInput.react";
 export { default as FormCheckboxInput } from "./forms/FormCheckboxInput.react";
